:recycle: move into effects/bases
