+++ skeleton-app/src/lib/classes/effects/bases/BaseEffect.ts
