+++ skeleton-app/src/lib/classes/effects/primitives/DrawEffect.ts
@@ -1,4 +1,4 @@
-import { BaseEffect } from "$lib/classes/effects/BaseEffect";
+import { BaseEffect } from "$lib/classes/effects/bases/BaseEffect";
 import { EffectType } from "$lib/types/effect";
 import type { DuelState } from "$lib/classes/DuelState";
 import type { EffectResult } from "$lib/types/effect";
